% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuRe.R
\name{parseStructure}
\alias{parseStructure}
\title{Parse a STRUCTURE output file to a data frame or matrix.}
\usage{
parseStructure(file, matrix = FALSE)
}
\arguments{
\item{file}{path to the STRUCTURE file.}

\item{matrix}{boolean; if true, returns only the matrix of inferred clusters.}
}
\description{
Parse a STRUCTURE output file to a data frame or matrix.
}
\keyword{STRUCTURE,}
\keyword{parse}
