% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuRe.R
\name{plot.PStructure}
\alias{plot.PStructure}
\title{Plot parsed STRUCTURE output.}
\usage{
\method{plot}{PStructure}(
  x,
  pop_names = NULL,
  col = NULL,
  yadjust = -0.03,
  pattern = "(^.{1}).*",
  order = "byname"
)
}
\arguments{
\item{x}{R object that inherits PStructure (parsed STRUCTURE).}

\item{pop_names}{optional; supply a vector of population names.}

\item{col}{optional; group colours. Default is greyscale.}

\item{yAdjust}{optional; adjust the y position of x axis labels.}
}
\description{
Plot parsed STRUCTURE output.
}
\keyword{STRUCTURE,}
\keyword{parse,}
\keyword{plot}
