% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuRe.R
\name{print.PStructure}
\alias{print.PStructure}
\title{Print parsed STRUCTURE output.}
\usage{
\method{print}{PStructure}(x)
}
\arguments{
\item{x}{R object that inherits PStructure (parsed STRUCTURE).}
}
\description{
Print parsed STRUCTURE output.
}
\keyword{STRUCTURE,}
\keyword{print}
