% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuRe.R
\name{plot.PStructure}
\alias{plot.PStructure}
\title{Plot parsed STRUCTURE output.}
\usage{
\method{plot}{PStructure}(
  x,
  pop_names = NULL,
  col = NULL,
  yadjust = -0.03,
  pattern = c("(^.{1}).*", "\\\\1"),
  order = "byname"
)
}
\arguments{
\item{x}{R object that inherits PStructure (parsed STRUCTURE).}

\item{pop_names}{optional; supply a vector of population names.}

\item{col}{optional; group colours. Default is greyscale.}

\item{pattern}{vector of length 2 describing the pattern and replacement for gsub() on the individual names to define groups.
the default regex takes the first letter from the vector of ID's.}

\item{order}{string of either byname (order by ID name), byq (order by q value of the first inferred cluster) or none, which orders the bars of the plot.}

\item{yAdjust}{optional; adjust the y position of x axis labels.}
}
\description{
Plot parsed STRUCTURE output.
}
\keyword{STRUCTURE,}
\keyword{parse,}
\keyword{plot}
